+++ packages/server_capability_core/CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Changed
 
+## [4.0.0-dev.1] - 2026-06-06
+
+### Changed
+
+- Align package version and hosted sibling dependency constraints with the Flutter MCP Toolkit breaking prerelease train.
+
 ## [0.1.0] - 2026-05-22
 
 ### Added